SQL Delight implementation.

+++ shared/entity/src/commonMain/kotlin/fetskovich/evgeny/entity/card/BankCard.kt
@@ -1,9 +1,9 @@
 package fetskovich.evgeny.entity.card
 
 data class BankCard(
-    val id: Int,
+    val id: Long,
     val bankCardType: BankCardType,
     val cardNumber: String,
     val expirationDate: Long,
-    val cvv: Int,
+    val cvv: Long,
 )
